# api-docs.json
{"openapi":"3.0.1","info":{"title":"Turtle Nest Monitoring Persistence DB API","description":"This projects establish the connection with the database, we don't have any logicimplemented in this project, since we only use this to save, consult and update information","version":"1.0"},"servers":[{"url":"http://localhost:8089/dbp","description":"Generated server url"}],"paths":{"/users/{id}":{"get":{"tags":["user-controllers"],"operationId":"getUserByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRes"}}}}}},"put":{"tags":["user-controllers"],"operationId":"updateUsersByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRes"}}}}}},"delete":{"tags":["user-controllers"],"operationId":"deleteUsersByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/sensors/{id}":{"put":{"tags":["sensors-controller"],"operationId":"updateSensorByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SensorsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SensorsRes"}}}}}},"delete":{"tags":["sensors-controller"],"operationId":"deleteSensorByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/records/{id}":{"get":{"tags":["records-controller"],"operationId":"getRecordById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecordsRes"}}}}}},"put":{"tags":["records-controller"],"operationId":"updateRecordByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecordsRes"}}}}}},"delete":{"tags":["records-controller"],"operationId":"deleteRecordByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/org/{id}":{"put":{"tags":["organizations-controller"],"operationId":"updateOrgByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationsRes"}}}}}},"delete":{"tags":["organizations-controller"],"operationId":"deleteOrgByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/nest/{id}":{"get":{"tags":["nests-controller"],"operationId":"getNestById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NestsRes"}}}}}},"put":{"tags":["nests-controller"],"operationId":"updateNestByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NestsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NestsRes"}}}}}},"delete":{"tags":["nests-controller"],"operationId":"deleteNestByID","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/users":{"get":{"tags":["user-controllers"],"operationId":"getUsers","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserRes"}}}}}},"post":{"tags":["user-controllers"],"operationId":"createUsers","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRes"}}}}}}},"/users/email":{"post":{"tags":["user-controllers"],"operationId":"getUserByEmail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRes"}}}}}}},"/sensors":{"get":{"tags":["sensors-controller"],"operationId":"getSensors","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageSensorsRes"}}}}}},"post":{"tags":["sensors-controller"],"operationId":"createSensor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SensorsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SensorsRes"}}}}}}},"/records":{"get":{"tags":["records-controller"],"operationId":"getRecords","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRecordsRes"}}}}}},"post":{"tags":["records-controller"],"operationId":"createRecord","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecordsRes"}}}}}}},"/org":{"get":{"tags":["organizations-controller"],"operationId":"getOrgs","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageOrganizationsRes"}}}}}},"post":{"tags":["organizations-controller"],"operationId":"createOrg","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrganizationsRes"}}}}}}},"/nest":{"get":{"tags":["nests-controller"],"operationId":"getNests","parameters":[{"name":"page","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageNestsRes"}}}}}},"post":{"tags":["nests-controller"],"operationId":"createNest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NestsReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NestsRes"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginReq"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/users/all":{"get":{"tags":["user-controllers"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRes"}}}}}}}},"/sensors/all":{"get":{"tags":["sensors-controller"],"operationId":"getAllSensors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SensorsRes"}}}}}}}},"/records/recent":{"get":{"tags":["records-controller"],"operationId":"getRecordsMostRecent","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRecordsRes"}}}}}}},"/records/nest/{nestID}":{"get":{"tags":["records-controller"],"operationId":"getRecordsByNestID","parameters":[{"name":"nestID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecordsRes"}}}}}}}},"/records/last/{nestID}":{"get":{"tags":["records-controller"],"operationId":"getLastRecordByNestID","parameters":[{"name":"nestID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecordsRes"}}}}}}},"/records/all":{"get":{"tags":["records-controller"],"operationId":"getAllRecords","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecordsRes"}}}}}}}},"/org/all":{"get":{"tags":["organizations-controller"],"operationId":"getAllOrganizations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrganizationsRes"}}}}}}}},"/nest/{assignedID}":{"get":{"tags":["nests-controller"],"operationId":"getNestByAssignedID","parameters":[{"name":"assignedID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NestsRes"}}}}}}},"/nest/sensor/{assignedID}":{"get":{"tags":["nests-controller"],"operationId":"getNestBySensorAssignedID","parameters":[{"name":"assignedID","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NestsRes"}}}}}}},"/nest/latest":{"get":{"tags":["nests-controller"],"operationId":"getLastsNests","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NestsRes"}}}}}}}},"/nest/all":{"get":{"tags":["nests-controller"],"operationId":"getAllNests","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NestsRes"}}}}}}}}},"components":{"schemas":{"UserReq":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"lastName":{"type":"string"},"isActive":{"type":"boolean"},"title":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"tempPassword":{"type":"string"},"isAdministrator":{"type":"boolean"}}},"UserRes":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"lastName":{"type":"string"},"isActive":{"type":"boolean"},"title":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"isAdministrator":{"type":"boolean"}}},"OrganizationsReq":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"dateActive":{"type":"string","format":"date-time"},"dateInactive":{"type":"string","format":"date-time"},"isActive":{"type":"boolean"},"contact":{"$ref":"#/components/schemas/UserReq"},"members":{"type":"array","items":{"$ref":"#/components/schemas/UserReq"}}}},"SensorsReq":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"assignedID":{"type":"string"},"softwareVersion":{"type":"string"},"hardwareVersion":{"type":"string"},"components":{"type":"string"},"temperatureError":{"type":"number","format":"float"},"moistureMinError":{"type":"number","format":"float"},"moistureMaxError":{"type":"number","format":"float"},"xerror":{"type":"number","format":"float"},"yerror":{"type":"number","format":"float"},"zerror":{"type":"number","format":"float"},"isInUse":{"type":"boolean"},"dateManufactured":{"type":"string","format":"date-time"},"organization":{"$ref":"#/components/schemas/OrganizationsReq"},"nrSumar":{"type":"number","format":"float"},"timing":{"type":"integer","format":"int32"}}},"OrganizationsRes":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"dateActive":{"type":"string","format":"date-time"},"dateInactive":{"type":"string","format":"date-time"},"isActive":{"type":"boolean"},"contact":{"$ref":"#/components/schemas/UserRes"},"members":{"type":"array","items":{"$ref":"#/components/schemas/UserRes"}}}},"SensorsRes":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"assignedID":{"type":"string"},"softwareVersion":{"type":"string"},"hardwareVersion":{"type":"string"},"components":{"type":"string"},"temperatureError":{"type":"number","format":"float"},"moistureMinError":{"type":"number","format":"float"},"moistureMaxError":{"type":"number","format":"float"},"xerror":{"type":"number","format":"float"},"yerror":{"type":"number","format":"float"},"zerror":{"type":"number","format":"float"},"isInUse":{"type":"boolean"},"dateManufactured":{"type":"string","format":"date-time"},"organization":{"$ref":"#/components/schemas/OrganizationsRes"},"nrSumar":{"type":"number","format":"float"},"timing":{"type":"integer","format":"int32"}}},"NestsReq":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"assignedID":{"type":"string"},"latitude":{"type":"string"},"longitude":{"type":"string"},"sensor":{"$ref":"#/components/schemas/SensorsReq"},"organization":{"$ref":"#/components/schemas/OrganizationsReq"},"responsible":{"$ref":"#/components/schemas/UserReq"},"isActive":{"type":"boolean"},"eggsQuantity":{"type":"integer","format":"int32"},"hatchlingsQuantity":{"type":"integer","format":"int32"},"depredatedEggsQuantity":{"type":"integer","format":"int32"},"infertileEggs":{"type":"integer","format":"int32"},"zone":{"type":"string"},"layingDate":{"type":"string","format":"date-time"},"collectionDate":{"type":"string","format":"date-time"},"monitoringStartDate":{"type":"string","format":"date-time"},"monitoringEndDate":{"type":"string","format":"date-time"},"firstHatchingRecordedDate":{"type":"string","format":"date-time"},"lastHatchingRecordedDate":{"type":"string","format":"date-time"},"nestCleaningDate":{"type":"string","format":"date-time"},"nestBehavior":{"type":"string"}},"description":"Nest object that needs to be added"},"RecordsReq":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nest":{"$ref":"#/components/schemas/NestsReq"},"samplingDateTime":{"type":"string","format":"date-time"},"temperature":{"type":"number","format":"float"},"humidityPercentage":{"type":"number","format":"float"},"energy":{"type":"number","format":"float"},"x":{"type":"number","format":"float"},"y":{"type":"number","format":"float"},"z":{"type":"number","format":"float"},"isTesting":{"type":"boolean"},"recordNumber":{"type":"integer","format":"int32"},"day":{"type":"number","format":"float"}},"description":"Record object that needs to be added"},"NestsRes":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"assignedID":{"type":"string"},"latitude":{"type":"string"},"longitude":{"type":"string"},"sensor":{"$ref":"#/components/schemas/SensorsRes"},"organization":{"$ref":"#/components/schemas/OrganizationsRes"},"responsible":{"$ref":"#/components/schemas/UserRes"},"isActive":{"type":"boolean"},"eggsQuantity":{"type":"integer","format":"int32"},"hatchlingsQuantity":{"type":"integer","format":"int32"},"depredatedEggsQuantity":{"type":"integer","format":"int32"},"infertileEggs":{"type":"integer","format":"int32"},"zone":{"type":"string"},"layingDate":{"type":"string","format":"date-time"},"collectionDate":{"type":"string","format":"date-time"},"monitoringStartDate":{"type":"string","format":"date-time"},"monitoringEndDate":{"type":"string","format":"date-time"},"firstHatchingRecordedDate":{"type":"string","format":"date-time"},"lastHatchingRecordedDate":{"type":"string","format":"date-time"},"nestCleaningDate":{"type":"string","format":"date-time"},"nestBehavior":{"type":"string"}}},"RecordsRes":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nest":{"$ref":"#/components/schemas/NestsRes"},"samplingDateTime":{"type":"string","format":"date-time"},"temperature":{"type":"number","format":"float"},"humidityPercentage":{"type":"number","format":"float"},"energy":{"type":"number","format":"float"},"x":{"type":"number","format":"float"},"y":{"type":"number","format":"float"},"z":{"type":"number","format":"float"},"isTesting":{"type":"boolean"},"recordNumber":{"type":"integer","format":"int32"},"day":{"type":"number","format":"float"}}},"LoginReq":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"PageUserRes":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserRes"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"PageSensorsRes":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/SensorsRes"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageRecordsRes":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RecordsRes"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageOrganizationsRes":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/OrganizationsRes"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageNestsRes":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/NestsRes"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}}}}